Refactoring balance type from `int` to `long` and changing `@AllArgsConstructor` to `@RequiredArgsConstructor` to comply with best practices

+++ src/main/java/com/payment/api/model/Account.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class Account {
 	private String id;
-	private Integer balance;
+	private long balance;
 }
